float(x,y) is deprecated, use decimal

+++ config/Migrations/20201014000220_MigrationCountries.php
@@ -72,7 +72,7 @@ class MigrationCountries extends AbstractMigration {
 				'null' => false,
 				'signed' => false,
 			])
-			->addColumn('lat', 'float', [
+			->addColumn('lat', 'decimal', [
 				'comment' => 'latitude',
 				'default' => null,
 				'null' => true,
